healthcare-agent-gke-autopilot/backend/requirements.txt
parlant>=3.2.0
google-genai==1.61.0
google-api-core==2.29.0
anthropic==0.78.0
pyjwt>=2.8.0
limits>=3.0.0
pymongo[srv]>=4.0.0 # use pymongo for self-hosted mongodb
redis>=5.0.0
